More refactoring. Fix an infinite migration loop caused by "dspace database migrate [version]" accidentally recalling DatabaseManager.initialize() (required moving MigrationUtils). Also add detailed warnings to this hidden migrate option.
